issues with blocks and function calls

+++ src/main/java/io/amberdata/domain/Block.java
@@ -1,5 +1,6 @@
 package io.amberdata.domain;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.Map;
 import java.util.Objects;
@@ -8,7 +9,7 @@ public final class Block implements BlockchainEntity {
     private BigInteger          number;
     private String              hash;
     private String              parentHash;
-    private BigInteger          gasUsed;
+    private BigDecimal          gasUsed;
     private Long                timestamp;
     private Integer             numTransactions;
     private Map<String, Object> optionalProperties;
@@ -47,11 +48,11 @@ public final class Block implements BlockchainEntity {
         this.parentHash = parentHash;
     }
 
-    public BigInteger getGasUsed () {
+    public BigDecimal getGasUsed () {
         return gasUsed;
     }
 
-    public void setGasUsed (BigInteger gasUsed) {
+    public void setGasUsed (BigDecimal gasUsed) {
         this.gasUsed = gasUsed;
     }
 
@@ -119,7 +120,7 @@ public final class Block implements BlockchainEntity {
         private BigInteger          number;
         private String              hash;
         private String              parentHash;
-        private BigInteger          gasUsed;
+        private BigDecimal          gasUsed;
         private Long                timestamp;
         private Integer             numTransactions;
         private Map<String, Object> optionalProperties;
@@ -139,7 +140,7 @@ public final class Block implements BlockchainEntity {
             return this;
         }
 
-        public Block.Builder gasUsed (BigInteger value) {
+        public Block.Builder gasUsed (BigDecimal value) {
             this.gasUsed = value;
             return this;
         }
